requirements.txt
langchain
langchain-huggingface
huggingface_hub
python-dotenv